requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

colorthief==0.2.1
langchain==0.1.8
langchain-community==0.0.21
langchain-openai==0.0.7
openai==1.12.0
opencv-python==4.9.0.80
python-dotenv==1.0.1
replicate==0.24.0
requests==2.31.0
webcolors==1.13
python-magic==0.4.15
tiktoken==0.6.0
unstructured==0.11.8
urllib3
qreader
pyzbar
boto3
fake-useragent
openpyxl
gradio
